Makes match ledger changes.

+++ agent-core/src/main/kotlin/org/knowledger/agent/core/ontologies/ledger/concepts/JBlockParams.kt
@@ -3,6 +3,6 @@ package org.knowledger.agent.core.ontologies.ledger.concepts
 import jade.content.Concept
 
 data class JBlockParams(
-    var blockMemSize: Long,
-    var blockLength: Long
+    var blockMemSize: Int,
+    var blockLength: Int
 ) : Concept
